merge new version

+++ engine/src/main/java/org/teiid/query/metadata/DatabaseStore.java
@@ -25,12 +25,16 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.locks.ReentrantLock;
 
 import org.teiid.api.exception.query.QueryResolverException;
+import org.teiid.connector.DataPlugin;
 import org.teiid.core.TeiidComponentException;
 import org.teiid.metadata.*;
 import org.teiid.metadata.Database.ResourceType;
 import org.teiid.metadata.Grant.Permission.Privilege;
+import org.teiid.metadata.Table.Type;
 import org.teiid.query.QueryPlugin;
 import org.teiid.query.parser.OptionsUtil;
+import org.teiid.query.parser.SQLParserUtil;
+import org.teiid.query.parser.SQLParserUtil.ParsedDataType;
 import org.teiid.query.resolver.util.ResolverUtil;
 import org.teiid.query.sql.symbol.GroupSymbol;
 import org.teiid.vdb.runtime.VDBKey;
@@ -55,7 +59,7 @@ public abstract class DatabaseStore {
     
     public void startEditing(boolean persist) {
         if (this.persist) {
-            throw new AssertionError();
+            throw new IllegalStateException();
         }
         this.lock.lock();
         this.persist = persist;
@@ -215,12 +219,16 @@ public abstract class DatabaseStore {
         }
     }
 
-    public void schemaDropped(String schemaName) {
+    public void schemaDropped(String schemaName, Boolean virtual) {
         if (!assertInEditMode(Mode.DATABASE_STRUCTURE)) {
             return;
         }
-        verifySchemaExists(schemaName);
-        assertGrant(Grant.Permission.Privilege.DROP, Database.ResourceType.SCHEMA, this.currentDatabase.getSchema(schemaName));
+        Schema s = verifySchemaExists(schemaName);
+        if (virtual != null && !virtual ^ s.isPhysical()) {
+            throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31273,
+                    QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31273, s.getName()));
+        }
+        assertGrant(Grant.Permission.Privilege.DROP, Database.ResourceType.SCHEMA, s);
         
         this.currentDatabase.removeSchema(schemaName);
 
@@ -228,45 +236,18 @@ public abstract class DatabaseStore {
             this.currentSchema = null;
         }
     }
-
-    protected Table verifyTableExists(String tableName) {        
-        verifyDatabaseExists();
-        Schema schema = this.getCurrentSchema();
-        
-        Table previous = schema.getTable(tableName);
-        
-        if (previous == null){
-            throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31237,
-                    QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31237, tableName, this.currentDatabase.getName()));        	
-        }
-        return previous;
-    }
     
-    
-    protected Procedure verifyProcedureExists(String procedureName) {        
-        verifyDatabaseExists();
-        Schema schema = this.getCurrentSchema();
-        
-        Procedure previous = schema.getProcedure(procedureName);
-        
-        if (previous == null){
-            throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31239,
-                    QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31239, procedureName, this.currentDatabase.getName()));        	
-        }
-        return previous;
-    }    
-
     protected FunctionMethod verifyFunctionExists(String functionName) {        
         verifyDatabaseExists();
         Schema schema = this.getCurrentSchema();
         
         for (FunctionMethod fm:schema.getFunctions().values()){
-        	if (fm.getName().equalsIgnoreCase(functionName)){
-        		return fm;
-        	}
+            if (fm.getName().equalsIgnoreCase(functionName)){
+                return fm;
+            }
         }
         throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31240,
-                QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31240, functionName, this.currentDatabase.getName()));        	
+                QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31240, functionName, this.currentDatabase.getName()));           
     }
     
     protected Schema verifySchemaExists(String schemaName) {        
@@ -419,44 +400,74 @@ public abstract class DatabaseStore {
         Schema s = getCurrentSchema();
         setUUID(s.getUUID(), table);
         
+        if (table.isVirtual() && table.getSelectTransformation() == null && table.getTableType() != Type.TemporaryTable) {
+            throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31272,
+                    QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31272, table.getFullName()));
+        }
+        
         s.addTable(table);
     }
     
-    public void setViewDefinition(final String tableName, final String definition, boolean updateFunctional) {
+    public void setViewDefinition(final String tableName, final String definition) {
         if (!assertInEditMode(Mode.SCHEMA)) {
             return;
         }
-        assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.TABLE, getCurrentSchema().getTable(tableName));
-        verifyTableExists(tableName).setSelectTransformation(definition);
+        Table table = (Table)getSchemaRecord(tableName, ResourceType.TABLE);
+        if (!table.isVirtual()) {
+            throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31238,
+                    QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31238, table.getFullName()));
+        }
+        assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.TABLE, table);
+        table.setSelectTransformation(definition);
     }
     
-    public void tableModified(Table table) {
+    public void modifyTableName(String name, Database.ResourceType type, String newName) {
         if (!assertInEditMode(Mode.SCHEMA)) {
             return;
         }
+        Table table = (Table)getSchemaRecord(name, type);
         assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.TABLE, table);
-        
-        verifyTableExists(table.getName());
-    }  
-    
-    public void modifyTableName(Table table, String newName) {
-        if (!assertInEditMode(Mode.SCHEMA)) {
-            return;
+        Schema s = table.getParent();
+        if (s.getTable(newName) != null) {
+            throw new DuplicateRecordException(DataPlugin.Event.TEIID60013, DataPlugin.Util.gs(DataPlugin.Event.TEIID60013, newName));
         }
-        assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.TABLE, table);
-        
-        verifyTableExists(table.getName());
+        s.getTables().remove(table.getName());
         table.setName(newName);
-    }     
+        s.getTables().put(newName, table);
+    }
     
-    public void tableDropped(String tableName) {
+    public void removeColumn(String objectName, Database.ResourceType type, String childName) {
         if (!assertInEditMode(Mode.SCHEMA)) {
             return;
         }
-        verifyTableExists(tableName);
-        assertGrant(Grant.Permission.Privilege.DROP, Database.ResourceType.TABLE, getCurrentSchema().getTable(tableName));
+        Table table = (Table)getSchemaRecord(objectName, type);
+        assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.TABLE, table);
         
-        Schema s = getCurrentSchema();
+        Column column = table.getColumnByName(childName);
+        if (column == null) {
+            throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31223,
+                QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31223, childName));      
+        }
+        table.removeColumn(column);
+    }
+    
+    public void tableDropped(String tableName, boolean globalTemp, boolean view) {
+        if (!assertInEditMode(Mode.SCHEMA)) {
+            return;
+        }
+        Table table = (Table)getSchemaRecord(tableName, ResourceType.TABLE);
+        assertGrant(Grant.Permission.Privilege.DROP, Database.ResourceType.TABLE, table);
+        
+        if (!(globalTemp ^ (table.getTableType() != Type.TemporaryTable))) {
+            throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31273,
+                    QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31273, table.getFullName()));
+        } 
+        if (view ^ table.isVirtual()) {
+            throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31273,
+                    QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31273, table.getFullName()));
+        }
+        
+        Schema s = table.getParent();
         s.removeTable(tableName);
     }
 
@@ -478,34 +489,37 @@ public abstract class DatabaseStore {
         s.addProcedure(procedure);
     }
     
-    public void setProcedureDefinition(final String procedureName, final String definition, boolean updateFunctional) {
+    public void setProcedureDefinition(final String procedureName, final String definition) {
         if (!assertInEditMode(Mode.SCHEMA)) {
             return;
         }
-        Procedure procedure = verifyProcedureExists(procedureName);
+        Procedure procedure = (Procedure)getSchemaRecord(procedureName, ResourceType.PROCEDURE);
         
-        assertGrant(Grant.Permission.Privilege.CREATE, Database.ResourceType.PROCEDURE, procedure);
+        assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.PROCEDURE, procedure);
+        
+        if (!procedure.isVirtual()) {
+            throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31238,
+                    QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31238, procedure.getFullName()));
+        }
         
         procedure.setQueryPlan(definition);
     }
     
-    public void procedureModified(Procedure procedure) {
+    public void procedureDropped(String procedureName, Boolean virtual) {
         if (!assertInEditMode(Mode.SCHEMA)) {
             return;
         }
-        verifyProcedureExists(procedure.getName());
-        assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.PROCEDURE, procedure);
-    }     
-
-    public void procedureDropped(String procedureName) {
-        if (!assertInEditMode(Mode.SCHEMA)) {
-            return;
+        Procedure procedure = (Procedure)getSchemaRecord(procedureName, ResourceType.PROCEDURE);
+        
+        if (virtual != null && virtual^procedure.isVirtual()) {
+            throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31273,
+                    QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31273, procedure.getFullName()));
         }
-        verifyProcedureExists(procedureName);
+        
 		assertGrant(Grant.Permission.Privilege.DROP, Database.ResourceType.PROCEDURE,
-				getCurrentSchema().getProcedure(procedureName));
+				procedure);
 
-        Schema s = getCurrentSchema();
+        Schema s = procedure.getParent();
         s.removeProcedure(procedureName);
     }
     
@@ -526,7 +540,7 @@ public abstract class DatabaseStore {
         s.addFunction(function);
     }
 
-    public void functionDropped(String functionName) {
+    public void functionDropped(String functionName, Boolean virtual) {
         if (!assertInEditMode(Mode.SCHEMA)) {
             return;
         }
@@ -542,7 +556,7 @@ public abstract class DatabaseStore {
         if (!assertInEditMode(Mode.SCHEMA)) {
             return;
         }
-        Table table = getCurrentSchema().getTable(tableName);
+        Table table = (Table)getSchemaRecord(tableName, ResourceType.TABLE);
         if (table == null) {
             throw new MetadataException(QueryPlugin.Util.getString("SQLParser.group_doesnot_exist", tableName)); //$NON-NLS-1$
         }
@@ -579,12 +593,11 @@ public abstract class DatabaseStore {
         }        
     }
     
-	public void enableTableTriggerPlan(final String tableName, final Table.TriggerEvent event, final boolean enable,
-			boolean updateFunctional) {
+	public void enableTableTriggerPlan(final String tableName, final Table.TriggerEvent event, final boolean enable) {
         if (!assertInEditMode(Mode.SCHEMA)) {
             return;
         }
-        Table table = getCurrentSchema().getTable(tableName);
+        Table table = (Table) getSchemaRecord(tableName, ResourceType.TABLE);
         if (table == null || !table.isVirtual()) {
             throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31244,
                     QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31244, tableName));                                                    
@@ -642,31 +655,10 @@ public abstract class DatabaseStore {
             case TABLE:
                 GroupSymbol gs = new GroupSymbol(name);
                 ResolverUtil.resolveGroup(gs, qmi);
-                Table t = (Table)gs.getMetadataID();
-                if (t == null) {
-                    throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31245,
-                            QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31245, name, getCurrentDatabase().getName(),
-                                    getCurrentSchema().getName()));
-                }
-                return t;
+                return (Table)gs.getMetadataID();
             case PROCEDURE:
                 StoredProcedureInfo sp = qmi.getStoredProcedureInfoForProcedure(name);
-                if (sp == null) {
-                    throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31213,
-                            QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31213, name, getCurrentSchema().getName(),
-                                    getCurrentDatabase().getName()));                
-                }
-                if (sp.getProcedureID() instanceof Procedure) {
-                    return (Procedure)sp.getProcedureID();
-                }                
-                return null;
-            case COLUMN:
-                Column c = qmi.getElementID(name);
-                if (c == null) {
-                    throw new org.teiid.metadata.MetadataException(QueryPlugin.Event.TEIID31223,
-                            QueryPlugin.Util.gs(QueryPlugin.Event.TEIID31223, name));                                    
-                }
-                return c;
+                return (Procedure)sp.getProcedureID();
             case DATABASE:
                 return getCurrentDatabase();
             case SCHEMA:
@@ -691,14 +683,13 @@ public abstract class DatabaseStore {
                 }
                 return dw;
             default:
-                break;
+                throw new AssertionError();
             }
         } catch (TeiidComponentException e) {
             throw new MetadataException(e);
         } catch (QueryResolverException e) {
             throw new MetadataException(e);
         }
-        return null;
     }
     
     protected TransformationMetadata getTransformationMetadata() {
@@ -843,9 +834,60 @@ public abstract class DatabaseStore {
         this.currentDatabase.revokeGrant(grant);
     }
     
-    public static MetadataFactory createMF(DatabaseStore events, Schema schema, boolean useSchema) {
+    public void renameBaseColumn(String objectName, Database.ResourceType type, String oldName, String newName) {
+        if (!assertInEditMode(Mode.SCHEMA)) {
+            return;
+        }
+        
+        MetadataFactory factory = DatabaseStore.createMF(this);
+        switch (type) {
+        case TABLE:
+            Table table = (Table)getSchemaRecord(objectName, type);
+            assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.TABLE, table);
+            factory.renameColumn(oldName, newName, table);
+            break;
+        case PROCEDURE:
+            Procedure proc = (Procedure)getSchemaRecord(objectName, type);
+            assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.PROCEDURE, proc);
+            factory.renameParameter(oldName, newName, proc);
+            break;
+        default:
+            throw new IllegalArgumentException("invalid type"); //$NON-NLS-1$
+        }
+    }
+    
+    public void alterBaseColumn(String objectName, Database.ResourceType type, String childName, ParsedDataType datatype, boolean autoIncrement, boolean notNull) {
+        MetadataFactory factory = DatabaseStore.createMF(this);
+        BaseColumn column = null;
+        if (type == Database.ResourceType.TABLE){
+            Table table = (Table)getSchemaRecord(objectName, type); 
+            assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.TABLE, table);
+            column = table.getColumnByName(childName);
+            if (column == null){
+                throw new ParseException(QueryPlugin.Util.getString("SQLParser.no_table_column_found", childName, table.getName())); //$NON-NLS-1$
+            }
+        } else {
+            Procedure proc = (Procedure)getSchemaRecord(objectName, type);
+            assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.PROCEDURE, proc);
+            column = proc.getParameterByName(childName);
+            if (column == null){
+                throw new ParseException(QueryPlugin.Util.getString("SQLParser.no_proc_column_found", childName, proc.getName())); //$NON-NLS-1$
+            }            
+        }
+        MetadataFactory.setDataType(datatype.getType(), column, factory.getDataTypes(), notNull);
+        SQLParserUtil.setTypeInfo(datatype, column);
+        if (notNull) {
+           column.setNullType(Column.NullType.No_Nulls);
+        }
+        if (type == Database.ResourceType.TABLE){
+            //must be called after setDataType as that will pull the defaults
+            ((Column)column).setAutoIncremented(autoIncrement);
+        }
+    }
+    
+    public static MetadataFactory createMF(DatabaseStore events, Schema schema, boolean useSchema, Properties modelProperties) {
         MetadataFactory mf = new MetadataFactory(events.getCurrentDatabase().getName(), events.getCurrentDatabase().getVersion(),
-                schema==null?"undefined":schema.getName(), events.getCurrentDatabase().getMetadataStore().getDatatypes(), new Properties(), null); //$NON-NLS-1$
+                schema==null?"undefined":schema.getName(), events.getCurrentDatabase().getMetadataStore().getDatatypes(), modelProperties, null); //$NON-NLS-1$
         Map<String, String> nss = events.getNameSpaces();
         for (String key:nss.keySet()) {
             mf.addNamespace(key, nss.get(key));    
@@ -858,7 +900,7 @@ public abstract class DatabaseStore {
     
     public static MetadataFactory createMF(DatabaseStore events) {
         //it's possible that the schema won't be set as we call this from the parsing context - statement will effectively be later ignored
-        return createMF(events, events.currentSchema, false);
+        return createMF(events, events.currentSchema, false, new Properties());
     }
     
     private void setUUID(String prefix, AbstractMetadataRecord record) {
@@ -918,5 +960,15 @@ public abstract class DatabaseStore {
 								record.getName(), this.commandContext.getUserName()));
 			}
 		}*/
-	}    
+	}
+
+    public Table getTableForCreateColumn(String objectName, ResourceType type) {
+        if (!assertInEditMode(Mode.SCHEMA)) {
+            return new Table(); //return a dummy table;
+        }
+        
+        Table table = (Table)getSchemaRecord(objectName, type); 
+        assertGrant(Grant.Permission.Privilege.ALTER, Database.ResourceType.TABLE, table);
+        return table;
+    }    
 }
